Incorrect title when searching
Added some template logic to the Header and Master Search Results modules to detect searching in a global context and

* Set the 'system page' flag, to avoid hundreds of 'about me' boxes on the sidebar
* Set a blog_override value (to 1)

Header will detect this and set the relevant strings to blog #1s, instead of giving the values from the first search result. (The <title> is being manually set to "$name search results", which may eventually need to change. This does not apply for search results on single blogs; see the distinction between:

http://blogs.perl.org/mt/mt-search.fcgi?IncludeBlogs=100&limit=20&search=database
and
http://blogs.perl.org/mt/mt-search.fcgi?limit=20&search=database

This addresses github issue39